requirements.txt
openai
fastapi
uvicorn
google-generativeai
langchain
python-dotenv
chromadb

bs4
feedparser
pandas